Restructured package hierarchy.
