+++ coralibre-sdk/sdk/src/androidTest/java/org/coralibre/android/sdk/internal/database/DatabaseTests.java
@@ -3,7 +3,6 @@ package org.coralibre.android.sdk.internal.database;
 import androidx.test.ext.junit.runners.AndroidJUnit4;
 import androidx.test.platform.app.InstrumentationRegistry;
 
-import org.coralibre.android.sdk.internal.crypto.ppcp.AssociatedMetadata;
 import org.coralibre.android.sdk.internal.crypto.ppcp.CryptoModule;
 import org.coralibre.android.sdk.internal.crypto.ppcp.ENNumber;
 import org.coralibre.android.sdk.internal.database.model.CapturedData;
